moves util-like functions to utils module
